# core/api/rest/openapi.yaml
...
openapi: 3.1.0
info:
  title: CypherPanel API
  description: >-
    Operator-facing control-plane API covering Phases 1–3: authentication and
    teams/roles, the server lifecycle, projects/environments, applications and
    the deploy pipeline, managed databases with S3 backups/restore, preview
    environments, notifications, and scheduled tasks. Authentication is a bearer
    session token from /api/v1/auth/login; the GitHub webhook authenticates by
    per-application HMAC instead. Every project-scoped route is authorized by
    team membership (a non-member sees 404, insufficient rank 403).


    Every response — success, error and SSE stream alike — carries an
    `X-Request-Id` header (`components/headers/RequestId`), and every JSON
    error body repeats it as `trace_id`. It is the value to quote in a bug
    report and the key to search for in `GET /api/v1/panel/logs`. Individual
    responses reference the header only where a generated client benefits;
    it is present on all of them.
  version: 0.3.0
servers:
  - url: /
...
paths:
...
  /api/v1/tokens:
    get:
      operationId: listTokens
      tags: [auth]
      summary: List the caller's personal access tokens (metadata only)
      description: >
        Requires an interactive session. Credential management is unreachable
        from a personal access token, whatever its abilities, so that a leaked
        token cannot mint itself a wider one.
      responses:
        "200":
          description: The caller's tokens, newest first. The secret is never returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiToken"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createToken
      tags: [auth]
      summary: Create a personal access token
      description: >
        Issues a scoped API token that authenticates as the calling user,
        inheriting that user's role and team memberships. The raw token is
        returned exactly once in this response and never again.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTokenRequest"
      responses:
        "201":
          description: The created token, including the raw secret (shown once).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateTokenResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/tokens/{id}:
    delete:
      operationId: deleteToken
      tags: [auth]
      summary: Revoke a personal access token
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Token revoked.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: Session token from /api/v1/auth/login.
  headers:
    RequestId:
      description: >-
        Correlation id for this request, minted by the API's outermost
        middleware and present on every response. It is repeated as `trace_id`
        in every JSON error body, and it is the key to search for in
        `GET /api/v1/panel/logs`. An inbound `X-Request-Id` is honoured only
        when the peer is a configured trusted proxy and the value is a short
        printable token; otherwise a fresh id is minted.
      schema: { type: string }
      example: trace_9f3a-11bd-4c02-7e51
  responses:
    BadRequest:
      description: Malformed or invalid request body.
      headers:
        X-Request-Id: { $ref: "#/components/headers/RequestId" }
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing, invalid, or expired session token.
      headers:
        X-Request-Id: { $ref: "#/components/headers/RequestId" }
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Forbidden:
      description: Authenticated but the caller's role is insufficient for this action.
      headers:
        X-Request-Id: { $ref: "#/components/headers/RequestId" }
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: >-
        No such resource — also returned when the caller is not a member of the
        owning team, so resource existence never leaks across tenants.
      headers:
        X-Request-Id: { $ref: "#/components/headers/RequestId" }
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable message; never contains secrets.
        trace_id:
          type: string
          description: >-
            The value of this response's `X-Request-Id` header, repeated in the
            body so a screenshot of a failure carries it. Paste it into a bug
            report or `GET /api/v1/panel/logs` to find the request. Present on
            every response served through the API's middleware.
          example: trace_9f3a-11bd-4c02-7e51
        retry_after_seconds:
          type: integer
          minimum: 1
          description: >-
            Present only on `429`: whole seconds until the throttle lifts, the
            same value as the `Retry-After` header. Count down against it rather
            than guessing.
...
    Ability:
      type: string
      enum: [read, write, deploy]
      description: >
        What a personal access token may do. `read` permits safe methods,
        `write` permits creating and changing resources, `deploy` permits
        triggering deploys and rollbacks. A token's authority is the
        intersection of its owner's role and its abilities.
    ApiToken:
      type: object
      required: [id, name, abilities, created_at]
      properties:
        id:
          type: string
          description: tok_… prefixed ID.
        name:
          type: string
          description: Human-readable label.
        abilities:
          type: array
          items:
            $ref: "#/components/schemas/Ability"
        last_used_at:
          type: string
          format: date-time
          description: When the token last authenticated a request; absent if never used.
        expires_at:
          type: string
          format: date-time
          description: When the token expires; absent if it never expires.
        created_at:
          type: string
          format: date-time
    CreateTokenRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          description: Human-readable label (1–100 characters).
        expires_in_days:
          type: integer
          description: Lifetime in days; 0 or omitted means the token never expires.
          minimum: 0
          maximum: 3650
        abilities:
          type: array
          description: >
            What this token may do. Omitted means full access (read, write,
            deploy); an explicit empty list is rejected.
          items:
            $ref: "#/components/schemas/Ability"
    CreateTokenResponse:
      allOf:
        - $ref: "#/components/schemas/ApiToken"
        - type: object
          required: [token]
          properties:
            token:
              type: string
              description: >
                The raw bearer token (cyp_… prefixed). Shown exactly once —
                store it now; it cannot be retrieved again.
...
security:
  - bearer: []